accounts 완료

+++ products/views.py
@@ -3,10 +3,11 @@ from django.http import JsonResponse, HttpResponse
 from django.core import serializers
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
+from rest_framework.pagination import PageNumberPagination
 from rest_framework import status
 from rest_framework.views import APIView
 from django.shortcuts import get_object_or_404
-from rest_framework.permissions import IsAuthenticated
+from rest_framework.permissions import IsAuthenticated, AllowAny
 from .models import Article, Comment
 from .serializers import (
     ArticleSerializer,
@@ -15,19 +16,31 @@ from .serializers import (
 )
 
 
-class ArticleListAPIView(APIView):
+class CustomPagination(PageNumberPagination):
+    page_size = 10
+    page_size_query_param = "page_size"
+    max_page_size = 100
 
-    permission_classes = [IsAuthenticated]
+
+class ArticleListAPIView(APIView):
+    permission_classes = [AllowAny]
 
     def get(self, request):
         articles = Article.objects.all()
+        paginator = CustomPagination()
+        paginated_articles = paginator.paginate_queryset(articles, request)
         serializer = ArticleSerializer(articles, many=True)
-        return Response(serializer.data)
+        return paginator.get_paginated_response(serializer.data)
 
     def post(self, request):
+        if not request.user.is_authenticated:
+            return Response(
+                {"detail": "로그인이 필요합니다."}, status=status.HTTP_403_FORBIDDEN
+            )
+
         serializer = ArticleSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
-            serializer.save()
+            serializer.save(author=request.user)
             return Response(serializer.data, status=201)
 
 
@@ -45,6 +58,11 @@ class ArticleDetailAPIView(APIView):
 
     def put(self, request, pk):
         article = self.get_object(pk)
+        if article.author != request.user:
+            return Response(
+                {"detail": "권한이 없습니다."}, status=status.HTTP_403_FORBIDDEN
+            )
+
         serializer = ArticleDetailSerializer(article, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -52,6 +70,11 @@ class ArticleDetailAPIView(APIView):
 
     def delete(self, request, pk):
         article = self.get_object(pk)
+        if article.author != request.user:
+            return Response(
+                {"detail": "권한이 없습니다."}, status=status.HTTP_403_FORBIDDEN
+            )
+
         article.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
